+++ packages/blaze-tx/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @blaze-cardano/tx
 
+## 0.4.0
+
+### Minor Changes
+
+- 578cfb8: Fix conditional when checking changeOutputIndex, where index 0 would result in a false false.
+
+## 0.3.6
+
+### Patch Changes
+
+- 1bfbe92: feat: Transaction cbor logging on eval error (@cjkoepke)
+  fix: Value merge was mutating the multiasset state of a value (thanks @cjkoepke, @Quantumplation)
+
 ## 0.3.5
 
 ### Patch Changes