+++ packages/blaze-wallet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @blaze-cardano/wallet
 
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies [578cfb8]
+  - @blaze-cardano/tx@0.4.0
+
+## 0.1.34
+
+### Patch Changes
+
+- Updated dependencies [1bfbe92]
+  - @blaze-cardano/tx@0.3.6
+
 ## 0.1.33
 
 ### Patch Changes